requirements.txt
streamlit
langchain-aws
boto3
python-dotenv